Break up the manager tests into sub classes
